+++ scorched/src/common/image/ImagePngFactory.cpp
@@ -1,5 +1,5 @@
 ////////////////////////////////////////////////////////////////////////////////
-//    Scorched3D (c) 2000-2011
+//    Scorched3D (c) 2000-2011, 2025
 //
 //    This file is part of Scorched3D.
 //
@@ -34,10 +34,10 @@ Image ImagePngFactory::loadFromFile(const char *filename, bool readalpha)
 	FILE *file = fopen(filename, "rb");
 	if (!file) return Image();
 
-	int read = 0;
+	size_t read = 0;
 	char buffer[256];
 	NetBuffer netBuffer;
-	while (read = (int) fread(buffer, 1, 256, file))
+	while ( ( read = fread(buffer, 1, 256, file) ) )
 	{
 		netBuffer.addDataToBuffer(buffer, read);
 	}
@@ -47,7 +47,8 @@ Image ImagePngFactory::loadFromFile(const char *filename, bool readalpha)
 	if (!result.getBits())
 	{
 		Logger::log(
-			S3D::formatStringBuffer("Failed to load PNG file \"%s\"", filename));
+			S3D::formatStringBuffer("Failed to load PNG file \"%s\"", filename)
+		);
 	}
 	return result;
 }
